+++ src/components/project/DisplayProjectCard.tsx
@@ -19,7 +19,7 @@ export const DisplayProjectCard = ({ project, hideClassId = false }: DisplayProj
     const normalizedSchedule = getNormalizedProjectSchedule(project)
 
     return (
-        <Link href={`/projects/${project.project_id}`}>
+        <Link href={`/projects/${project.project_id}`} prefetch={false}>
             <Card className="h-full overflow-hidden transition-all hover:shadow-lg hover:border-primary/50 flex flex-col bg-white py-0">
                 {project.image_url ? (
                     <div className="aspect-[4/3] w-full overflow-hidden bg-muted relative">
